+++ cmd/gen_errors/gen.go
@@ -5,6 +5,7 @@ package main
 import (
 	"fmt"
 	"log"
+	"os"
 	"path/filepath"
 
 	"github.com/unlimited-budget-ecommerce/errz"
@@ -35,4 +36,26 @@ func main() {
 	}
 
 	fmt.Println("Generated", outputFile)
+}
+
+func projectRoot() (string, error) {
+	dir, err := os.Getwd()
+	if err != nil {
+		return "", fmt.Errorf("cannot get working directory: %w", err)
+	}
+
+	for !fileExists(filepath.Join(dir, "go.mod")) && dir != "/" {
+		dir = filepath.Dir(dir)
+	}
+
+	if dir == "/" {
+		return "", fmt.Errorf("project root not found (no go.mod)")
+	}
+
+	return dir, nil
+}
+
+func fileExists(path string) bool {
+	_, err := os.Stat(path)
+	return err == nil
 }
